Module 46 Modulation architecture
