+++ app/build.gradle
@@ -1,8 +1,12 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
+plugins {
+    id 'com.android.application'
+//apply plugin: 'kotlin-android'
 //apply plugin: 'kotlin-android-extensions'
-apply plugin: 'kotlin-parcelize'
-apply plugin: 'kotlin-kapt'
+//    id 'com.android.library'
+    id 'org.jetbrains.kotlin.android'
+    id 'kotlin-parcelize'
+    id 'kotlin-kapt'
+}
 
 android {
     compileSdk 33
@@ -10,10 +14,10 @@ android {
 
     defaultConfig {
         applicationId "com.example.filmsSearch"
-        minSdk 21
+        minSdk 30
         targetSdk 31
-        versionCode 1
-        versionName "1.0"
+        //versionCode 1
+        //versionName "1.0"
 
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
@@ -41,7 +45,7 @@ android {
 apply from: "$project.rootDir/versions.gradle"
 dependencies {
     implementation fileTree(dir: "libs", include: ["*.jar"])
-    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.8.22'
+    implementation 'org.jetbrains.kotlin:kotlin-stdlib-jdk8:1.9.0'
 //    implementation 'org.jetbrains.kotlin:kotlin-gradle-plugin:1.8.20'
     implementation 'androidx.core:core-ktx:1.10.1'
     implementation 'androidx.appcompat:appcompat:1.6.1'
@@ -49,6 +53,7 @@ dependencies {
     implementation 'androidx.cardview:cardview:1.0.0'
     implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation 'com.google.android.gms:play-services-wearable:18.0.0'
+    implementation project(path: ':remote_module')
 
     //Tests
     testImplementation 'junit:junit:4.13.2'
@@ -82,20 +87,20 @@ dependencies {
     implementation "androidx.room:room-runtime:2.5.2"
     kapt "androidx.room:room-compiler:2.5.2"
     //EventBus
-    implementation 'org.greenrobot:eventbus:3.2.0'
+    implementation 'org.greenrobot:eventbus:3.3.1'
     //ViewModels (val mainActivityViewModel: MainActivityViewModel by viewModels())
     //in Fragments by activityViewModels()
     implementation "androidx.fragment:fragment-ktx:1.6.1"
     //Coroutines
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.4.2'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.7.3'
     //Coroutines for ROOM
     implementation "androidx.room:room-ktx:2.5.2"
     //Room
     implementation "androidx.room:room-rxjava3:2.5.2"
     //RxJava
-    implementation 'io.reactivex.rxjava3:rxandroid:3.0.0'
-    implementation 'io.reactivex.rxjava3:rxjava:3.0.10'
+    implementation 'io.reactivex.rxjava3:rxandroid:3.0.2'
+    implementation 'io.reactivex.rxjava3:rxjava:3.1.6'
     implementation 'io.reactivex.rxjava3:rxkotlin:3.0.1'
     //RxJava for retrofit
     implementation "com.github.akarnokd:rxjava3-retrofit-adapter:3.0.0"
