--- app/src/main/java/com/example/filmsSearch/data/Entity/LogoX.kt
@@ -1,9 +1,0 @@
-package com.example.filmsSearch.data.Entity
-
-
-import com.google.gson.annotations.SerializedName
-
-data class LogoX(
-    @SerializedName("url")
-    val url: String
-)
