+++ app/src/main/java/com/example/filmsSearch/di/AppComponent.kt
@@ -2,25 +2,24 @@ package com.example.filmsSearch.di
 
 import com.example.filmsSearch.di.modules.DatabaseModule
 import com.example.filmsSearch.di.modules.DomainModule
-import com.example.filmsSearch.di.modules.RemoteModule
 import com.example.filmsSearch.view.viewmodel.DetailsFragmentViewModel
 import com.example.filmsSearch.view.viewmodel.FavoritesFragmentViewModel
 import com.example.filmsSearch.view.viewmodel.HomeFragmentViewModel
 import com.example.filmsSearch.view.viewmodel.SettingsFragmentViewModel
+import com.example.remote_module.RemoteProvider
 import dagger.Component
 import javax.inject.Singleton
 
 @Singleton
 @Component(
-    //Внедряем все модули, нужные для этого компонента
+    dependencies = [RemoteProvider::class],
     modules = [
-        RemoteModule::class,
         DatabaseModule::class,
         DomainModule::class
     ]
 )
+
 interface AppComponent {
-    //метод для того, чтобы появилась возможность внедрять зависимости в HomeFragmentViewModel
     fun inject(homeFragmentViewModel: HomeFragmentViewModel)
 
     fun inject(settingsFragmentViewModel: SettingsFragmentViewModel)
