+++ app/src/main/java/com/example/filmsSearch/utils/Converter.kt
@@ -1,20 +1,19 @@
 package com.example.filmsSearch.utils
 
-import com.example.filmsSearch.data.Entity.Doc
 import com.example.filmsSearch.data.Entity.Film
 
 object Converter {
-    fun convertApiListToDtoList(list: List<Doc>?): List<Film> {
+    fun convertApiListToDtoList(list: List<com.example.remote_module.entity.Doc>?): List<Film> {
         val result = mutableListOf<Film>()
         list?.forEach {
             result.add(
                 Film(
-                title = it.name,
-                poster = it.poster.url,
-                description = it.description,
-                rating = it.rating.kp,
-                isInFavorites = false
-            )
+                    title = it.name,
+                    poster = it.poster.url,
+                    description = it.description,
+                    rating = it.rating.kp,
+                    isInFavorites = false
+                )
             )
         }
         return result
