+++ remote_module/src/main/java/com/example/remote_module/RemoteComponent.kt
@@ -1,0 +1,8 @@
+package com.example.remote_module
+
+import dagger.Component
+import javax.inject.Singleton
+
+@Singleton
+@Component(modules = [RemoteModule::class])
+interface RemoteComponent: RemoteProvider
