+++ remote_module/src/main/java/com/example/remote_module/RemoteProvider.kt
@@ -1,0 +1,5 @@
+package com.example.remote_module
+
+interface RemoteProvider {
+    fun provideRemote(): TmdbApi
+}
