+++ remote_module/src/main/java/com/example/remote_module/TmdbApi.kt
@@ -1,6 +1,6 @@
-package com.example.filmsSearch.data
+package com.example.remote_module
 
-import com.example.filmsSearch.data.Entity.TmdbResultsDto
+import com.example.remote_module.entity.TmdbResultsDto
 import io.reactivex.rxjava3.core.Observable
 import retrofit2.Call
 import retrofit2.http.GET
