+++ remote_module/src/main/java/com/example/remote_module/entity/ApiConstants.kt
@@ -1,4 +1,4 @@
-package com.example.filmsSearch.data
+package com.example.remote_module.entity
 
 object ApiConstants {
     const val BASE_URL = "https://api.kinopoisk.dev/"
