+++ remote_module/src/main/java/com/example/remote_module/entity/Logo.kt
@@ -1,0 +1,9 @@
+package com.example.remote_module.entity
+
+
+import com.google.gson.annotations.SerializedName
+
+data class Logo(
+    @SerializedName("url")
+    val url: Any?
+)
