+++ remote_module/src/main/java/com/example/remote_module/entity/TmdbResultsDto.kt
@@ -1,4 +1,4 @@
-package com.example.filmsSearch.data.Entity
+package com.example.remote_module.entity
 
 
 import com.google.gson.annotations.SerializedName
