+++ versions.gradle
@@ -1,5 +1,5 @@
 ext {
-    daggerVersion = '2.46.1'
+    daggerVersion = '2.47'
     dagger = "com.google.dagger:dagger:$daggerVersion"
     daggerCompiler = "com.google.dagger:dagger-compiler:$daggerVersion"
 }
